Added tests for retry, timeout and recover task type
