+++ test/com/linkedin/parseq/TestTask.java
@@ -97,71 +97,6 @@ public class TestTask extends AbstractTaskTest {
     testOnFailureWhenCancelled(3);
   }
 
-  @Test
-  public void testFusionTaskType() {
-    Task<Integer> task = getSuccessTask().map(String::length);
-    runAndWait("fusionTaskType", task);
-    assertEquals(task.getShallowTrace().getTaskType(), TaskType.FUSION.getName());
-  }
-
-  @Test
-  public void testBlockingTaskType() {
-    TestingExecutorService es = new TestingExecutorService(Executors.newSingleThreadExecutor());
-    try {
-      Task<String> task = Task.blocking(() -> "blocking task", es);
-      runAndWait("blockingTaskType", task);
-      assertEquals(task.getShallowTrace().getTaskType(), TaskType.BLOCKING.getName());
-    } finally {
-      es.shutdown();
-    }
-  }
-
-  @Test
-  public void testShareableTaskType() {
-    Task<Integer> value = Task.value(10);
-    Task<Integer> shareableTask = value.shareable();
-    assertEquals(shareableTask.getShallowTrace().getTaskType(), TaskType.SHAREABLE.getName());
-  }
-
-  @Test
-  public void testFlatMapTaskType() {
-    Task<String> task = Task.value("Welcome");
-    Task<String> flatMap = task.flatMap("+earth", s -> Task.callable(() -> s + " on earth!"));
-    runAndWait("flatMapTaskType", flatMap);
-    assertEquals(flatMap.getShallowTrace().getTaskType(), TaskType.FLATTEN.getName());
-  }
-
-  @Test
-  public void testWithTimeoutTaskType() {
-    Task<?> taskWithTimeout = Task.value("test").withTimeout(50, TimeUnit.MILLISECONDS);
-
-    runAndWait("taskWithTimeoutTaskType", taskWithTimeout);
-    Assert.assertEquals(taskWithTimeout.getShallowTrace().getTaskType(), TaskType.WITH_TIMEOUT.getName());
-  }
-
-  @Test
-  public void testWithSideEffectTaskType() {
-    Task<?> taskWithSideEffect = Task.value("value1Task", "value1").withSideEffect("delayed sideEffect",
-        v -> delayedValue("value2", 100, TimeUnit.MILLISECONDS));
-
-    runAndWait("taskWithSideEffectTaskType", taskWithSideEffect);
-    Assert.assertEquals(taskWithSideEffect.getShallowTrace().getTaskType(), TaskType.WITH_SIDE_EFFECT.getName());
-  }
-
-  @Test
-  public void testWithRetryTaskType() {
-    Task<String> task = Task.withRetryPolicy(RetryPolicy.attempts(1, 0), attempt -> Task.value("successful attempt " + attempt));
-    runAndWait(task);
-    assertEquals(task.getShallowTrace().getTaskType(), TaskType.WITH_RETRY.getName());
-  }
-
-  @Test
-  public void testWithRecoverTaskType() {
-    Task<String> task = getFailureTask().recoverWith(e -> Task.value("recoveryTask"));
-    runAndWait(task);
-    assertEquals(task.getShallowTrace().getTaskType(), TaskType.WITH_RECOVER.getName());
-  }
-
   @Override
   Task<String> getSuccessTask() {
     return Task.async("success", () -> Promises.value(TASK_VALUE));
